Improve clipping in pyramid creation (#54)
Since the source 16-bit images are rescaled to 8-bits for the viewer, the shortest intensities range that should be used is such that there are 256 levels of gray in the final image. In that way, if there is low signal, then the original intensities are maintained in the viewer and the user can see finer gradients. 

Also, the behavior for clipping with user-defined upper clipping thresholds is improved (background is set to 0.001 percentile instead of 0).